+++ screens/ExploreEvents/index.ts
@@ -1,1 +1,2 @@
 export * from "./ExploreView"
+export * from "./useExploreEvents"
